pyproject.toml
[project]
name = "repoaiproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "celery>=5.6.3",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "pygithub>=2.9.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "svix>=1.90.0",
    "uvicorn>=0.44.0",
]
